CPKAFKA-2260: prioritize tiering partitions with lowest lag in bytes (#309)
Archiving the partitions with the minimum lag first behaves better when initially catching up for newly tiered topics, as it minimizes tiering data that will be deleted by retention shortly after. It also maximizes tiering of segments that are most likely to still be in the page cache after being written.