CPKAFKA-2763: Performance Tests: Add tail consumer test type and RunUntiLEnd scheduling (#317)
* Extend performance test scheduler to support RunUntilEnd scheduling

RunUntilEnd means that a job (A) is scheduled to run until the end of another job B. In such a case, job A should not have a defined duration.
We do not support chained RunUntil definitions. e.g job B can run until job A. job C *cannot* run until job B. job C can run until job A

ScenarioContext is the context of this scenario - all the tests what are being parsed. The idea is to have it be passed around tests such that tests can get information from other tests by name

* Add "run_until" definition to schedule JSON specification

You can now schedule a test with an unknown duration time to run until the end of another test. If multiple tests are passed to "run_until", the one with the latest end time is picked. It is mandatory that the test scheduled to run until another test does not have any explicit duration specified

* Add a TailConsumer test type

The TailConsumer test spawns Trogdor ConsumeBench tasks with the goal of raiding from the tail of the log.