CONFLUENT: Add hasProducerState field to TierObjectMetadata (#399)
* Add hasProducerState field to TierObjectMetadata.

This format change is aimed at supporting producer state propagation in
tiered storage. When a leader starts up, it will begin archiving from the
beginning of a topic. Because Kafka currently deletes all but the last two
producer state snapshots, this means that there will be some segments which
have no accompanying producer state to upload.

We wish to avoid a situation where a leader deletes local log segments after
tiering, preventing a follower from being able to materialize it's own
producer state. To do this, we will write out a ~hasProducerState~ flag along
with each each TierObjectMetadata entry, signifying if the archived segment
has an accompanying producer state snapshot. We will then restrict the leader
from deleting any local segments until there is at least one local segment
which is eligible for deletion (lies outside of the hotset) which has an
accompanying producer state snapshot file, signified by
~hasProducerState=true~.

A future patch will disable the deletion logic for producer state snapshots
entirely, keeping them around for the lifetime of the segment.

These changes should avoid the situation where a follower cannot either
materialize producer state, or retrieve a snapshot of producer state from
tiered storage.

The ~hasProducerState~ field is not strictly necessary, as it would also be
possible to issue ~GET~ requests against the TierObjectStore and interpret
failed requests as ~hasProducerState=false~. We add it because it may help
with debugging in the future, and allows us to use existing codepaths for
state materialization and determining if a local segment is eligible for
deletion.

* CONFLUENT: Prefix TierObjectMetadata boolean with `has`.