CONFLUENT: encode tier topic IDs and message ID as base64 in S3 keys (#421)
This change is primarily to improve human readability of topic IDs and messageIDs in S3 keys by base64 encoding the UUIDs contained in the S3 key. It also minimizes the S3 key length as the base64 representation is shorter. Updates the toString methods for the AbstractTierMetadata classes to allow for log correlation with the keys residing in S3.