CPKAFKA-2252: Tiered Storage transactions (#332)
Introduce support for `read_committed` from tiered storage.

The approach taken here is to upload all aborted transactions between a
segments base offset and last offset. It's also expected that the
TierArchiver is unable to archive past the last stable offset. Because we
only upload segments which encompass segment ranges lower than or up to the
last stable offset, it's guaranteed that we have the full set of aborted
transactions for a segment at upload time. This allows us to upload all
aborted transactions for the segment range at the time we tier the segment.

The TierFetcher can then fetch 1 file per segment to determine all of the
aborted transactions within that segment before returning them to the Consumer.