Add partition throughput metrics by client id (#462)
We want to be able to expose partition level throughput broken down by
client id. This adds a new throughput meter with an additional tag.
The existing partition level throughput rate is left as is so as to not
affect percentile statistics.

This change also refactors tenant metrics to make it easier to add
arbitrary tags to the metrics context based on any metadata available
within the request context.

We also add a special tag to prevent those metrics from getting exposed
via JMX, since scraping a large number of those can have performance impacts.