CONFLUENT: shutdown tier components in order of startup (#520)
This fixes a couple of issues in our startup/shutdown order:

Several components such as the TierArchiver depend on the TierObjectStore, and yet the TierObjectStore is shutdown first.

The TierMetadataManager is shutdown prior to the TierTopicManager. Doing so will clear the TierPartitionState map, meaning that the TierTopicManager may read tier state topic entries, ignore them, and then commit the tier.offsets file at a point past those entries, even though they are required for correct state materialization.

We now shutdown each component in the reverse order of startup.