CONFLUENT: Change the archiver ForkJoin pool to a fixed thread pool. (#518)
The TierArchiver previously used a ForkJoin pool, which proved a bit unpredictable in practice as to when it would spawn a new thread. This changes the TierArchiver to use a fixed size thread pool.

Set default Archiver thread count to 4.

Increase TierArchiverStateTest timeouts to 30 seconds, to ensure we have sufficient time to wait on future execution to prevent CI flakes.