CONFLUENT: fix resumption of materialization after restart prior to ConfluentLeaderAndIsrRequest (#536)
There is a bug in the TierMetadataManger where the broker will resume materialization after a restart, but the TierPartitionState will not be in the metadataById map, even though the TierPartitionState was previously ONLINE and had been assigned a TopicIdPartition. This causes metadata to miss materialization causing TierPartitionState divergence.

Adds tests for TierTopicManager resume logic and listener clearing.