CONFLUENT: reduce TierArchiver default thread count (#542)
At a thread count of 4, the TierArchiver can tier 100MB+/sec. This can have a negative impact on produce rate, consume rate, and end to end latency. This PR reduces the thread count to 2, which should still be capable of catching on lag, as the tier rate should still be approximately 2x the expected produce rate for our CCloud clusters.