METRICS-275 New Telemetry MetricsReporter with OpenCensus serialization (#505)
New Telemetry MetricsReporter with OpenCensus serialization

New implementation of the MetricsReporter interface,
`KafkaServerMetricsReporter`, with components to collect metrics on CPU and
volumes and metrics from KafkaMetrics and YammerMetrics. These components are
intended to be reusable by additional CP products, and there is a modular
design that allows for collecting of certain metrics and exporting to
additional systems. The initial implementation includes a KafkaExporter for
producing metrics to a Kafka topic.