ST-1857: Fix CP version in BasicCollector and FullCollector (#569)
This stopped working when we migrated this code to Gradle in the
confluentinc/kafka repository. Previously, the version.properties
file was populated by Maven. Rely on post-processing of the kafka
version instead.

Without this change, the basic collector is basically broken as the
server will reject messages that have ${project.version} as the
version. A similar issue may occur for the full collector.

Reviewers: Manikumar Reddy <manikumar.reddy@gmail.com>