CPKAFKA-3106: TierTopicManager producer.partitionsFor may timeout on startup (#544)
There is an issue in the TierTopicManager startup, where the initial partitionsFor may exceed the producer's max.block.ms and we do not retry it. Instead, the thread will exit, blocking the ability to tier data. This PR switches to using the TierAdminManager and AdminClient to also check the number of partitions on startup, and retry if the request fails.

This also fixes a second issue, where the TierTopicManager may hang on shutdown for up to 120 seconds, as the TierAdminClient timeout is 120 seconds by default. As part of the above change, I have reduced this timeout to 30 seconds. It will instead use the usual TierTopicManager retry logic if the request fails. This fixes a lot of TierFeatureAdminIntegrationTest flakiness, where the tests would timeout after two minutes, due to starting up and shutting down the broker in quick succession.

Tests: TierTopicAdminTest tests tier state topic creation behavior, and retry logic.