CPKAFKA-3262: TierTopicManager sees incorrect __confluent-tier-state partition count on startup (#576)
When we find an invalid tier state topic partition count, this is now considered a retriable exception. This is required because on startup the broker may have partial metadata about the topic and partition counts, so we cannot treat this error as fatal until https://issues.apache.org/jira/browse/KAFKA-8480 is fixed. This ensures that the TierTopicManager worker thread doesn't die, but it does not let it become ready either, so this condition will not lead us to produce tier metadata to an incorrect partition.

Testing:
The TierTopicAdmin test checks whether a TierTopicIncorrectPartitionCountException is correctly returned.