CPKAFKA-3067: Enable acls by default for MultiTenantAuthorizer, update error (#487)
We are enabling ACLs for CCP. This PR updates the error message when ACLs are disabled. It also enables ACLs by default when MultiTenantAuthorizer is configured since it longer makes sense to disable by default when all environments expect it to be enabled. Note the change in default doesn't impact CCloud since we explicitly configure max.acls=0 for CCP at the moment, so we can merge this PR before CCloud is updated.

Reviewers: Manikumar Reddy <manikumar.reddy@gmail.com>