CONFLUENT: Handle mismatch between partition count for tier topic (#601)
We do not support changing the partition count of the tier topic. For various reasons, the current logic is to wait until the actual partition count matches the configured count. The downside of such an approach is that when the default partition count changes, the TierTopicManager would keep looping, waiting for the partition count to match the updated default.

This PR allows us to ignore the configured partition count when tier topic has already been created. We instead log a warning and continue using the topic as it exists.