CPKAFKA-2720: ISR should exclude observer brokers (#545)
CPKAFKA-2720: ISR should exclude observer brokers

At a high-level this commit introduces the following changes:

1. Limit which brokers are allowed to join the ISR. Only the brokers that
meet the following criteria are allowed to join the IRS.
  a. The broker is the preferred replica. The preferred replica is the
  first replica in the topic partition assignment.

  b. The broker matches the preferred replica constraint configuration.

  c. The broker matches the leader rack/attributes as specified by the
  insync eligible configuration.

2. Exposes a new "CaughtUpReplicas" metric which returns the number of
replicas which are insync with the leader. Note that this is always
greater then or equal to the size of the ISR.

3. Redefine the "UnderMinIsr" metric to be true when the size of the ISR
is less than the size of replicas that are allowed to join the ISR.

4. Related to bullet 1., this change adds two topic only configuration
options: insync.eligible.replica.attributes and
preferred.replica.constraints.