CPKAFKA-3411: Allow all replicas to join the ISR if the leader matches the observers constraints (#625)
CPKAFKA-3411: Allow all replicas to join the ISR if the leader matches the observers constraints

Because we want to support "dynamic" replicas and observers, we need to
restrict the set of valid topic placement configuration. The validator
will not allow observers constraints if there are no replicas
constraints.

Based on the above restriction we also modify matchesReplicas so that
any broker that matches the replicas constraints when replicas constraints is not set.

Finally, the leader should allow all replicas to join the ISR if the
leader matches the observers constraints. This is needed in case unclean
leader election is performed and an observers is elect as leader.

Note: I think we can simplify this logic and relax the topic placement
validator and the matching logic once we implement "static" replicas and
observers in the controller, leader, followers and observers.