CPKAFKA-2719: Honor replica placement constraints when assigning replica (#615)
* CPKAFKA-2719: Honor replica placement constraints when assigning replica

This change updates replica assignment code to honor the constraints
specified using TopicPlacement specification. The constraint is
specified using a JSON that gets converted to a TopicPlacement object.
Then the constraint in the object is read and brokers are places in
different set based on the constraint. After that for each set of
brokers replica assignment is done. Finally, union of all assignments
for a partition is returned.

When no replica constraint json is specified, the code falls back on
original behavior.

The change also adds a new "replica-placement" command line option to
the TopicCommand. If this option is specified, the json file contains
the distribution of replica to different racks/zones/regions. In this
case we enforce that "replication-factor" option is not provided.