CPKAFKA-3405: Avoid queueing deletion events when topic deletion is being completed (#624)
To support tiered storage, we added an additional step in the deletion state machine that enqueues an event to the controller queue to complete topic deletion, once all replicas have been successfully deleted. Every time we "resume" topic deletion, we could end up queueing duplicate completion events to the controller queue, which could cause increased pressure on the controller. This patch fixes the issue by ensuring duplicate completion events are not queued. We also removed this extra state transition if the topic is not tiered.

Also improves logging in TopicDeletionManager.