Add `kafka_broker_version` label to all metrics. (#638)
Refactors the KafkaServerMetricsReporter to use the OpenCensus Resource
entity for global labels.  The Resource object is set in the Context
and used to build all Metric objects. A new MetricBuilderFacade class
is added to replace MetricUtil for construction of Metrics with flat labels
and single timeseries data points.  The MetricBuilderFacade is constructed
via the Context to enforce that the Resource is always set on the Metric.

Resource labels are duplicated on each metric timeseries for backwards compatibility.