MINOR. Update kakfa-topics.sh execution to use security configurations (#690)
A previous commit updated code for `kafka-topics.sh' command to use the
new --bootstrap-server option. That exposed a few issues that this
commit fixes:

1. Some command line flags are not compatible with --bootstrap-server
2. SASL related broker deployements were not working.

For first issue, following switches are not compatible with
--bootstrap-server option:
1. --delete-config
2. --if-exists
3. --if-not-exists

We only use `--if-not-exists' in systest, so this change checks for this
option and switches to --zookeeper if this option is present.

For the second issue, this change now get the client security
configuration from the SecurityConfig object and passes that as
--command-conifg option. It also specifies the JASS and Kerberos
configuration as environment variables.

Tested by running both observers_test and replication_test to confirm
that test cases pass for both SASL enabled and plain Kafka brokers.