CPKAFKA-3538: Make the empty string the default value for placement constraints (#701)
With a default value of null some clients (C3) can send the string "null" for a value. This causes the JSON decoder to return null and results in a null pointer exception when validating the JSON object. To avoid this, change the default value to the empty string "" and change the topic placement parser to return the an empty topic placement for this case.