MINOR. Skip --bootstrap-server config when testing with server not th… (#702)
* MINOR. Skip --bootstrap-server config when testing with server not that do not support it

We have AK system tests and Muckrake ones that use older version of
Kafka servers to test. Control Center also have many of these type of
tests. The change to use --bootstrap-server for create topic thus broke
these tests.

This change checks the version of the Kafka server that it talks to and
then switches to older --zookeeper option if it is talking to a kafka
server that is pre 2.3.

--command-config option isn't present for older kafka versions either. This is
also not needed when zookeeper option is present. So skipping it for that
case and now the code behaves same as it used to before introducing
--bootstrap-server option.

Tested by running `system-test-ce-kafka-branch-builder` job and made sure that all tests impacted by this change pass.