Support pulling private dependencies into ce-kafka (#679)
This change pulls in the telemetry-api library (built from the observability
"schroedinger" monorepo) into the ce-kafka build. The telemetry-api builds on
top of OpenCensus metrics with additional
protobuf definitions. The telemetry-api is used by the ce-metrics module.

Even though opencensus-proto is transitively pulled in by telemetry-api we
prefer a direct dependency since ce-metrics code directly references the
opencensus-proto classes.

The confluent-nexus maven repo is added as a repository to the ce-metrics
project, using the existing mavenUsername and mavenPassword properties.
I decided against the original approach of using m2/settings.xml to avoid
introducing the complexity of "mirror" configurations. For example, a side
effect of using m2/settings.xml (with the plugin) is that the URL of the
mavenCentral() repository is actually replaced with nexus.confluent.io. This is
a more substantial change than I want to introduce with this PR.

The Jenkins withCredentials task is used to populate the mavenUsername and
mavenPassword properties as environment variables from the secret stored in
Jenkins. Using a Jenkins secret provides reuse with other jobs and also avoids
the need to store nexus credentials in an unencrypted config file in Jenkins.

To remove all dependencies on the Jenkins config file script, the mavenUrl
property used by the uploadArchives task is now set explicitly via the
command-line (instead of the previous approach of using gradle init script).

The Makefile is modified to copy the users local gradle properties into the
docker build context (project directory). The Dockerfile then copies this into
GRADLE_USER_HOME within the build image.

METRICS-454

See 1-pager for additional background:

https://confluentinc.atlassian.net/wiki/spaces/OBSERVE/pages/919942907/Dependency+management+for+Observability+Java+libraries