[QE-35] Create Sustained Produce/Consume Workloads for Soak CLI (#723)
* [QE-35] Create Sustained Produce/Consume Workloads for Soak CLI

This is a modification to the existing Produce and Consume workloads within the Soak Go library.  This originally started out as separate methods with separate requirements, but I believe I've found a way to merge all the requirements into one.

Previous iterations can be found here: https://github.com/confluentinc/ce-kafka/pull/459/commits/8f950c33a8faa228ceb1935dda2715cbc77ecded

This PR adds several features to the existing tests:

- Ability to specify MessagesPerSecond for Consumer tasks.
  - This allows us the ability to run throttled consumer tasks.  The default still remains unthrottled.

- Ability to run Producer/Consumer tasks at differing numbers than the Trogdor Nodes.
  - This gives us the ability to guarantee tests are the same when running with different-sized trogdor clusters.  The default still remains equal tasks across all available nodes.

- Ability to run with staggered start for all tasks.
  - This allows us to ramp up load rather than opening the floodgates and shocking the system.  The default still remains unstaggered.

- Ability for a Produces tasks to run against a specific topic.
  - This allows us to point at any topic and run, rather than just the single name we had before.  The default still remains the original topic name.

- Ability to change the message padding for Producer tasks.
  - This allows us to have complete control over the message size.  This is only applicable if the MessageSize parameter is also modified.  The default still remains 100.

- Ability to specify a consumer group for Consumer tasks.
  - Before we were pigeon-holed into a generated consumer group, this way we can have explicit control over the consumer group.  The default still remains the generated consumer group.

This passes the go lint and compiles, however I have not yet run tests using it.  I am getting this out for review because it is late on Friday and I wanted to get all of the above written down before the weekend.