CPKAFKA-3583; Add configurable request log sampling (#715)
This patch implements dynamically configurable request sampling. The intent is to use this to get more insight into current usage and to facilitate debugging. Specifically this is intended to help in scenarios where we have a sudden flood of requests to one of the clusters. The rate of sampling is controlled using a fixed interval configuration. No more than one sample will be logged for each interval, but a given interval may not contain any samples. This patch allows a default sampling interval which applies to all APIs and custom intervals targeting specific APIs. The idea is that we could turn this on by default at a conservative rate to start getting more insight into usage. We also have the option for more targeted adjustment for specific events.

Reviewers: Guozhang Wang <wangguoz@gmail.com>