Add Telemetry MetricsReporter Http Exporter stats (#774)
Collect metrics from the Telemetry Client

Adds a new interface MetricsCollectorProvider, that can be
implemented by a Metrics Exporter. If implemented, then we
register that exporter as a collector. This PR adds an
implementation for the HttpExporter.

Also, adds tests for MetricsReporter startup. These ensure that
various configuration errors result in an exception, and that we
can startup given a base set of properties. Several fields of
the MetricsReporter were exposed for testing in order to test
those items.

Refs: METRICS-530, METRICS-681