CPKAFKA-3707: Adds API for fetching the replica status of partitions
Adds replicaStatus() call to the KafkaAdminClient and command
kafka-replica-status.sh for querying the status of partitions'
replicas (leader, followers, and observers). Includes information
such as the lag time behind the follower that for a replica,
whether the leader considers it to be caught up, and current log
start/end offsets.