KSTORAGE-274: Handle concurrent DeleteRecordsRequest when archiving segment (#793)
When the log start offset is incremented (on DeleteRecords for example), the producer snapshot files are deleted right away but the segments still exist and are garbage collected in the background. When the archiver runs into an exception, we check the segment map to validate if the segment was deleted but this check is insufficient if the log start offset was incremented past it, causing the snapshot file to be deleted for example.

This patch fixes this issue by factoring in the log start offset to determine if a segment was deleted when uploading. It also improves the logic to determine the next tierable segment by accounting for the bumped log start offset.