CONFLUENT: Allow holes in tag ids (#815)
* CONFLUENT: Allow holes in tag ids

This change relaxes the restriction that tag ids should be contiguous
and should start with 0. This is to allows us to add fields with a
higher tag number that doesn't conflict with the lower tag numbers that
can be added in AK.

Tested by adding new tags starting with id 10000 and confirmed that the
compilation error goes away.