CPKAFKA-3830: Include observer information in MetadataResponse (#831)
* CPKAFKA-3830: Include observer information in MetadataResponse

This change updates the code to include observer information in metadata
response. The information is part of `UpdateMetadataRequest` api which
gets saved in the MetadataCache. The MetadataCache code is also updated
to populate observers when creating `MetadataResponse.PartitionMetadata`
as well as when creating a ClusterMetadata object.

Tested by deploying along with fix to `describeTopics` api and made sure
that the observer information is appearing in the metadata response and
as part of TopicCommand `--describe` option.