MINOR: Use LeaderAndIsr instead of ConfluentLeaderAndIsr when tiered is enabled (#843)
We only do this if the inter-broker protocol is 2.4-IV1 so that we maintain
compatibility with clusters currently running tiered in prod.

More concretely, we add an optional tagged topicId field to `LeaderAndIsrRequest`
and we use it if tiered is enabled and IBP >= 2.4. The expectation is that
CP 5.4 customers will use this. I would suggest requiring IBP >= 2.4
for tiered support in the 5.4.x branch (not included in this PR). Once all Cloud
clusters are running with IBP 2.4, we can delete `ConfluentLeaderAndIsrRequest`
and respective conditional logic from the code. Once topicIds are added
to Apache Kafka, we can deprecate the proprietary optional tagged field.

Adjusted `LeaderAndIsrRequestTest` to check topic id serialization/deserialization
and added a test to `KafkaApisTest` that verifies we handle a mix of set and unset
topic ids correctly.

Reviewers: Lucas Bradstreet <lucas@confluent.io>