CPKAFKA-3783: Set observers in CreatePartitons call (#879)
* CPKAFKA-3783: Set observers in CreatePartitons call

When create partitions call is made, a replica assigment for new
partition is provided by the user. This change checks this replica
assignment and then get observers from this list. This set of observers
then get saved as topic configuration.

If no replica assignment is provided by client in the call, then we need
to use topic placement constraint to create the replica assignment and
save it. This change uses same code that is used when creating a new
topic.

Added new test cases to test newly added code.