KSTORAGE-252: update partition state endOffset on deleteInitiate (#820)
Issue:
Operations such as retention and deleteRecords() can result in segments to be deleted from tiered storage. The deletionTask determines the "segment" to be deleted and then updates the LogStartOffset to segment.endOffset + 1
However for performance reasons the checkpoint LogStartOffset is executed via async job or on clean shutdown.
https://confluentinc.atlassian.net/browse/KSTORAGE-252
This is a case where leader called deleteInitiate for retention reasons and potentially had a unclean shutdown without logstartoffset checkpoint happen (we don't have the logs to absolutely confirm, but based on code review this is highly likely)
On an unclean leader election the broker may attempt to re-upload the same segment again if it deleteInitiate was called for the segment, checkpoint of logStartOffset may or may not have executed. The change essentially updates the endOffset to track the deleteInitiate and then allows new object for the same segment to be uploaded.

The PR implements the following :-
EndOffset tracking for tiered partition
The end offset normally should get updated when segments are successfully uploaded. In order to track the endOffset correctly we are now going to record this as part of state file and not depending on the segment metadata to derive this information
Rollout Challenge : We are going to add endOffset for the new broker code. So where the state file is materialized the endOffset will get saved in the file when the next flush occurs. We are proposing that for the edge case (discussed above) we will ignore the case where endOffsets cannot be evaluated correctly, for these cases we will manually mitigate by disabling tiered storage deleting the state files so that it then rematerializes with the correct endOffset.

Background:
In the current design we use the states to evaluate the endOffset individually on brokers (leader and followers). We have a bug where the endOffset diverge from leader and follower in the case where the leader ends up re-uploading the deleted segment when it comes back up from an unclean shutdown where its logStartOffset was not checkpoint correctly, this happens because the deleteInitiate message gets fenced (As epoch changes) but on followers deleteInitiate does not get fenced and does not approve uploading a duplicate segment (that was previously deleted) and therefore the divergence in endOffset between the follower and the leader causing archiver thread to throw and exception and therefore stopping archiving on the followers.

We rejected the approach to distinguish between the uploads that are fenced from deletes that were fenced, but this ends up introducing two news states and making this more complex.