KSTORAGE-232: use ListOffsetRequest rather than TierListOffsetRequest when tiering enabled (#934)
Tiered storage currently requires TierListOffsetRequest, which is used to fetch the start and end offset on a broker's local disk for replication purposes. We will now use the timestamp field of ListOffsetRequest to achieve the same behavior. This will allow us to avoid using TierListOffsetRequest in 5.4, and to also delete TierListOffsetRequest from our codebase after CCloud is running IBP 2.4 across all our clusters.