CPKAFKA-4034: Retry on UnsupportedVersionException for license topic create (#943)
CreateTopics request was added in 0.10.1.0. When doing a rolling upgrade from older versions to Confluent Server, AdminClient requests to create license topic may fail if the request is sent to an older node that doesn't support CreateTopics. This commit performs retry for UnsupportedVersionException similar to InvalidReplicationFactorException when brokers are starting up.

Reviewers: Ismael Juma