KSTORAGE-309: invalid handling of full fetch sessions for tiered fetches (#941)
Full fetch sessions require responses to contain all partitions included in the request or else the consumer will reject the response as invalid. Normally this will now show up in consumer behavior because an incremental fetch session is generally created if the FetchSessionCache has enough slots or can evict another entry.

This change ensures that TierLogReadResult(s) return entries when one of the partitions is in error state.

We now test that all partitions are accounted for in the ReplicaManager.fetchMessages callback when one partition is in error and the other partition is tiered.