CPKAFKA-4086 Fix configuration in reassign_partitions_test.py (#985)
Explicitly set the broker configs needed for this test before the tiering config is set.

Also, improve the way this test waits for its setup precondition. Rather than sleeping for a precise amount of time, we poll the broker for the log offsets to see when the old segment has been deleted.