METRICS-848 disable telemetry reporter for anything but brokers (#981)
this disables telemetry reporter in the event that either cluster id is or broker id is null. This should prevent clients from accidentally instantiating the telemetry reporter in most common cases.

It is still possible for a degenerate client config to have a broker.id setting, in which case this check would not prevent the problem, unfortunately there isn't a better way to detect a broker today.