KSTORAGE-323: invalid handling of SegmentDeleteInitiate for fenced segments (#1005)
SegmentDeleteInitiate messages for fenced segments incorrectly remove unfenced segments from the validSegments state in the FileTierPartitionState. This prevents them from being fetched or deleted. This is a non-destructive error, as the partition state file itself is still valid and is not mutated to remove the unfenced segment.

We now check that the segment being removed from validOffsets matches the objectId that we expect before removing it. I have added a unit test that tests this path, and another test that tests the handling of reopened TierPartitionState files when fenced segments are deleted. The latter did not find a bug but does improve our test coverage.