KSTORAGE-318: Trogdor record batch verifier auto.offset.reset=none (#980)
The old record batch/offset verifier had a problem where offset validation would fail if the consumer does not fetch quickly enough for a partition and encounters an offset out of range error. This change sets auto.offset.reset to none for the consumer when offset validation is in effect. By doing so, we can now detect offset out of range errors and reset the offset we are validating (the next offset will not match the expected offset). Upon hitting an offset out of range error, we now manually seek the partition to the beginning.