CPKAFKA-4116: Add new metric for response send I/O time (#1042)
Adds a new metric to track actual I/O time for sending responses:
kafka.network:type=RequestMetrics,name=ResponseSendIoTimeMs,request=<apiKey>

This is similar to the existing metric ResponseSendTimeMs which includes the full wallclock time from the start of send until send completes, including any time waiting for the channel to be ready for write.

The new metric ResponseSendIoTimeMs is also a yammer histogram and uses the same unit as ResponseSendTimeMs, so the difference between the two metrics indicates the time when the channel was not ready for write, e,g, because the client was not polling. The value of this metric is calculated using the existing network thread timestamps for each channel that are always tracked on the broker for managing request quotas. So the value is accurate and doesn't add any additional overhead.

Reviewers: Ismael Juma