CONFLUENT. Fix constructor chaining arguments (#1047)
* CONFLUENT. Fix constructor chaining arguments

When adding a new C'tor for taking set of observers, existing C'tor was
chained to call the new C'tor with default observer argument. But the
method mistakenly passed a default argument for the `offlineReplicas`
variable too. Its a simple typo that this change fixes.

The C'tor that takes observers is now private. A new builder method can
now be used to create a new PartitionInfo object.

Ran unit tests and they pass. There is no known code that directly uses
the broken C'tor, but as its a public api someone may be using it
outside kafka repository.