[CNKAF-153] Support for pod-namespace tag in the storage probe stats. (#1039)
Add support for a "namespace" tag in the storage-probe stats exported by the Kafka K8s storage probe, in order to allow for faster divination of which cluster/pod might be having problems.

The initial implementation of the storage probe tagged metrics with the pod-name, and has a general set of other tags including the host name, but not the namespace. Without this information, we can identify that a given host is having storage difficulties but have to do a reverse mapping
to figure out which pod that is.

By attaching the pod-namespace that we're reporting for as a tag, it ?should be? easy to identify which unique pod is having problem with high storage latencies/errors.