KSTORAGE-349: Synchronize log deletion and tier partition state flush (#1062)
On log deletion, we rename the partition directory to -deleted. For tiered storage, we have a periodic task to flush the tier partition state file which involves making a copy of the state file to the partition directory. As there is no synchronization between log deletion and the flush, it could be that the flush attempts to write to a non-existent file path which causes the broker to crash.

This patch synchronizes log deletion and tier partition state flush to fix this issue. While this is not ideal as it may block log creation until we are able to flush state for all partitions, it may be good enough for now.

Also added a unit test to verify the changes. Without the patch, the test failed almost every time. With the patch, the test ran successfully 100 times.