KSTORAGE-289: support upgrading from CP and confluent.tier.feature = true (#1035)
Adds support to the upgrade system test to allow upgrading from Confluent Platform to master. Also adds a "from_tiered_storage" parameterization to allow upgrading from a version of kafka with tiered storage turned on to another version with tiered storage turned on in both cases. This is not activate by this commit, because it will require a version with tiered storage to upgrade from, and this not the case for any released version of Confluent Platform.